+++ Cargo.toml
@@ -5,13 +5,12 @@ lto = true
 
 [workspace]
 members = [
-    "lib/flv",
-    "lib/rtmp",
-    "lib/fmp4",
+    "lib/codec",
     "lib/golomb",
     "lib/balance",
     "lib/transport",
     "lib/configure",
+    "lib/multiplexer",
     "exchange",
     "publish",
     "pull"
