+++ lib/codec/src/lib.rs
@@ -2,10 +2,15 @@
 //!
 //! All codec instances must implement the `Codec` and` Default` traits.
 //! Provide an abstract interface for the TcpSocket instance to use.
+//! 
+
+#[macro_use]
+extern crate lazy_static;
 
 mod rtmp;
 
-pub use self::rtmp::Rtmp;
+pub use rtmp::Rtmp;
+
 use bytes::BytesMut;
 use transport::Flag;
