--- lib/fmp4/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "fmp4"
-version = "0.1.0"
-authors = ["Mr.Panda <xivistudios@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-bytes = "0.5.4"
-
-[dependencies.flv]
-path = "../flv"
