+++ lib/multiplexer/Cargo.toml
@@ -1,15 +1,11 @@
 [package]
-name = "rtmp"
+name = "multiplexer"
 version = "0.1.0"
 authors = ["Mr.Panda <xivistudios@gmail.com>"]
 edition = "2018"
 
 [dependencies]
 bytes = "0.5.4"
-rml_rtmp = "0.3.2"
-rml_amf0 = "0.1.1"
-lazy_static = "1.4.0"
 
-
-[dependencies.transport]
-path = "../transport"
+[dependencies.golomb]
+path = "../golomb"
