+++ publish/Cargo.toml
@@ -8,8 +8,8 @@ edition = "2018"
 bytes = "0.5.4"
 futures = "0.3"
 
-[dependencies.rtmp]
-path = "../lib/rtmp"
+[dependencies.codec]
+path = "../lib/codec"
 
 [dependencies.transport]
 path = "../lib/transport"
