+++ pull/Cargo.toml
@@ -10,8 +10,8 @@ bytes = "0.5.4"
 futures = "0.3"
 http = "0.2"
 
-[dependencies.flv]
-path = "../lib/flv"
+[dependencies.multiplexer]
+path = "../lib/multiplexer"
 
 [dependencies.transport]
 path = "../lib/transport"
