+++ pull/src/server/socket.rs
@@ -8,7 +8,7 @@ use std::sync::Arc;
 use std::task::{Context, Poll};
 use std::{error::Error, pin::Pin};
 use transport::{Flag, Payload};
-use flv::{Tag, Flv};
+use multiplexer::{Tag, Flv, Header};
 use tungstenite::{
     handshake::server::Request, 
     handshake::server::Response, 
@@ -60,7 +60,7 @@ impl Socket {
     /// TODO：这是一个固定的头，
     /// 后期可以优化为常量避免重复分配.
     fn packet_header() -> Message {
-        let flv_packet = Flv::encode_header(flv::Header::Full);
+        let flv_packet = Flv::encode_header(Header::Full);
         Message::Binary(flv_packet.to_vec())
     }
